requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
blas=1.0=mkl
ca-certificates=2021.1.19=h06a4308_0
certifi=2020.12.5=py38h578d9bd_1
cudatoolkit=10.2.89=hfd86e86_1
cycler=0.10.0=pypi_0
decorator=4.4.2=pypi_0
freetype=2.10.4=h5ab3b9f_0
fvcore=0.1.3.post20210204=pyhd8ed1ab_0
imageio=2.9.0=pypi_0
intel-openmp=2020.2=254
iopath=0.1.3.post20210120=py38
jpeg=9b=h024ee3a_2
kiwisolver=1.3.1=pypi_0
lcms2=2.11=h396b838_0
ld_impl_linux-64=2.33.1=h53a641e_7
libedit=3.1.20191231=h14c3975_1
libffi=3.3=he6710b0_2
libgcc-ng=9.1.0=hdf63c60_0
libpng=1.6.37=hbc83047_0
libstdcxx-ng=9.1.0=hdf63c60_0
libtiff=4.1.0=h2733197_1
libuv=1.40.0=h7b6447c_0
lz4-c=1.9.3=h2531618_0
matplotlib=3.3.4=pypi_0
mkl=2020.2=256
mkl-service=2.3.0=py38he904b0f_0
mkl_fft=1.2.0=py38h23d657b_0
mkl_random=1.1.1=py38h0573a6f_0
ncurses=6.2=he6710b0_1
networkx=2.5=pypi_0
ninja=1.10.2=py38hff7bd54_0
numpy=1.19.2=py38h54aff64_0
numpy-base=1.19.2=py38hfa32c7d_0
olefile=0.46=py_0
opencv-python=4.5.1.48=pypi_0
openssl=1.1.1i=h27cfd23_0
pillow=8.1.0=py38he98fc37_0
pip=20.3.3=py38h06a4308_0
plotly=4.14.3=pypi_0
portalocker=1.7.0=py38h578d9bd_1
pyparsing=2.4.7=pypi_0
python=3.8.5=h7579374_1
python-dateutil=2.8.1=pypi_0
python_abi=3.8=1_cp38
pytorch=1.7.1=py3.8_cuda10.2.89_cudnn7.6.5_0
pytorch3d=0.4.0=py38_cu102_pyt171
pywavelets=1.1.1=pypi_0
pyyaml=5.3.1=py38h8df0ef7_1
readline=8.1=h27cfd23_0
retrying=1.3.3=pypi_0
scikit-image=0.18.1=pypi_0
scipy=1.6.0=pypi_0
setuptools=52.0.0=py38h06a4308_0
six=1.15.0=py38h06a4308_0
sqlite=3.33.0=h62c20be_0
tabulate=0.8.7=pyh9f0ad1d_0
termcolor=1.1.0=py_2
tifffile=2021.2.1=pypi_0
tk=8.6.10=hbc83047_0
torchvision=0.8.2=py38_cu102
tqdm=4.56.1=pyhd8ed1ab_0
typing_extensions=3.7.4.3=pyha847dfd_0
wheel=0.36.2=pyhd3eb1b0_0
xz=5.2.5=h7b6447c_0
yacs=0.1.6=py_0
yaml=0.2.5=h516909a_0
zlib=1.2.11=h7b6447c_3
zstd=1.4.5=h9ceee32_0
